requirements.txt
Django~=4.0
api~=0.0.7
djangorestframework~=3.13.1
djangorestframework_simplejwt~=5.0.0